Create SUBNETSearchDHCPvsKidControlMACUnknownTOGenericUser
This is my script that automatically adds a Unknown Client Machine to Kid Control in a Specified subnet if the device has not been assigned already. We use this script to block addresses and stop random mac generation from devices forcing the client to turn off that feature to use the network.  It uses DHCP to search and scans Kid Control mac addresses to validate each machine and machines that do not validate the process are added to a generic account that is restricted. You must create that account.